Changed Double values to BigDecimal

+++ src/main/java/com/miniconomy/commercial_bank_service/dto/TransactionRequest.java
@@ -1,11 +1,13 @@
 package com.miniconomy.commercial_bank_service.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class TransactionRequest {
     private Long debitAccountId;
-    private Double transactionAmount;
+    private BigDecimal transactionAmount;
     private String creditRef;
     private String debitRef;
 }
