+++ src/main/java/com/miniconomy/commercial_bank_service/response/DebitOrderResponse.java
@@ -1,5 +1,7 @@
 package com.miniconomy.commercial_bank_service.response;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -10,7 +12,7 @@ public class DebitOrderResponse
   private Long creditAccountId;
   private Long debitAccountId;
   private String debitOrderCreatedDate;
-  private Double debitOrderAmount;
+  private BigDecimal debitOrderAmount;
   private String debitOrderReceiverRef;
   private String debitOrderSenderRef;
 }
